Move home page to authentication app
