+++ property_web/.gitignore
@@ -6,3 +6,6 @@ __pycache__/
 
 # Ignore file .env
 .env
+
+# Ignore media folder
+media/
