+++ property_web/authentication/templates/home.html
@@ -1,37 +1,40 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>{% block title %}HomeHaven{% endblock %}</title>
-</head>
-<body>
-    <h1>HOMEHAVEN</h1>
+<!-- home_page.html -->
+{% extends 'base.html' %}
 
-    {% if user.is_authenticated %}
-        <p>Hi, {{ user.username }}!</p>
-        <form action="{% url 'logout' %}" method="post">
-            {% csrf_token %}
-            <button type="submit">Logout</button>
-        </form>
-        <form action="{% url 'change_password' %}" method="get">
-            <button type="submit">Change Password</button>
-        </form>
-        <form action="{% url 'create_post' %}" method="get">
-            <button type="submit">Create Post</button>
-        </form>
-        <form action="{% url 'user_detail' user_id=user.id %}" method="get">
-            <button type="submit">Information user</button>
-        </form>
-    {% else %}
-        <form action="{% url 'login' %}" method="get">
-            <button type="submit">Login/Register</button>
-        </form>
-    {% endif %}
-    
-    <div class="content">
-        {% block content %}
-        {% endblock %}
-    </div>
-</body>
-</html>
+{% block content %}
+    {% load static %}
+    <link rel="stylesheet" href="{% static 'css/list_post_home.css' %}">
+
+    <h2 class="subheading">Hot Posts</h2>
+    <ul class="post-list">
+        {% for post in hot_posts %}
+            <li class="post-item">
+                <a href="{% url 'post_detail' post_id=post.id %}">
+                    {% if post.image %}
+                        <img class="post-image-ava" src="{{ post.image }}" alt="{{ post.title }} Image">
+                    {% endif %}
+                    <span>{{ post.title }}</span>
+                    <span>{{ post.price }}</span>
+                    <span>{{ post.address }}</span>
+                </a>
+            </li>
+        {% endfor %}
+    </ul>
+
+    <h2 class="subheading">Normal Posts</h2>
+    <ul class="post-list">
+        {% for post in normal_posts %}
+            <li class="post-item">
+                <a href="{% url 'post_detail' post_id=post.id %}">
+                    {% if post.image %}
+                        <img class="post-image-ava" src="{{ post.image }}" alt="{{ post.title }} Image">
+                    {% endif %}
+                    <span>{{ post.title }}</span>
+                    <span>{{ post.price }}</span>
+                    <span>{{ post.address }}</span>
+                </a>
+            </li>
+        {% endfor %}
+    </ul>
+    <a href="{% url 'list_posts' %}" class="btn">Xem thêm</a>
+{% endblock %}
