+++ property_web/authentication/urls.py
@@ -4,13 +4,13 @@ from .views import (
     user_login,
     change_password,
     user_logout,
-    home
+    home_page
     )
 
 urlpatterns = [
+    path('', home_page, name='home'),
     path('login/', user_login, name='login'),
     path('signup/', user_signup, name='signup'),
     path('changepassword/', change_password, name='change_password'),
     path('logout/', user_logout, name='logout'),
-    path('home/', home, name='home')
 ]
