+++ property_web/posts/templates/list_posts.html
@@ -1,5 +1,5 @@
 <!-- post_list.html -->
-{% extends 'home.html' %}
+{% extends 'base.html' %}
 
 {% block content %}
     {% load static %}
