--- property_web/posts/templates/list_posts_home.html
@@ -1,39 +1,0 @@
-<!-- post_list.html -->
-{% extends 'home.html' %}
-
-{% block content %}
-    {% load static %}
-    <link rel="stylesheet" href="{% static 'css/list_post_home.css' %}">
-
-    <h2 class="subheading">Hot Posts</h2>
-    <ul class="post-list">
-        {% for post in hot_posts %}
-            <li class="post-item">
-                <a href="{% url 'post_detail' post_id=post.id %}">
-                    {% if post.image %}
-                        <img class="post-image-ava" src="{{ post.image }}" alt="{{ post.title }} Image">
-                    {% endif %}
-                    <span>{{ post.title }}</span>
-                    <span>{{ post.price }}</span>
-                    <span>{{ post.address }}</span>
-                </a>
-            </li>
-        {% endfor %}
-    </ul>
-
-    <h2 class="subheading">Normal Posts</h2>
-    <ul class="post-list">
-        {% for post in normal_posts %}
-            <li class="post-item">
-                <a href="{% url 'post_detail' post_id=post.id %}">
-                    {% if post.image %}
-                        <img class="post-image-ava" src="{{ post.image }}" alt="{{ post.title }} Image">
-                    {% endif %}
-                    <span>{{ post.title }}</span>
-                    <span>{{ post.price }}</span>
-                    <span>{{ post.address }}</span>
-                </a>
-            </li>
-        {% endfor %}
-    </ul>
-{% endblock %}
