+++ property_web/posts/templates/post_detail.html
@@ -1,5 +1,5 @@
 <!-- post_detail.html -->
-{% extends 'home.html' %}
+{% extends 'base.html' %}
 
 {% block content %}
     <div class="post-detail">
