+++ property_web/property_web/urls.py
@@ -21,7 +21,7 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('user/', include('users.urls')),
-    path('auth/', include('authentication.urls')),
+    path('home/', include('authentication.urls')),
     path("admin/", admin.site.urls),
     path('posts/', include('posts.urls'))
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
